011220-14:20
Simplified the implementation for responses. Responses now function on their own without explicit definitions for conversations or stages, instead these are implemented implicitly.

Prefabs in TestingGrounds scene have been updated to comprise of two varaints of response - NPC and others. NPC response prefabs are non-interactable and are set to red on Awake in Response.cs.