160121-19:00
Added menu buttons which enable the player to back out of a conversation, engage one, or quit the application altogether.

Canvas objects appear and disappear appropriately with the use of a "toggled" boolean which is used by the START/BACK menu button.

Enabled the GameManager.cs class to manage timer events rather than the HUDManager.cs class.