190121-18:00
Added white background image to intro scene to block out any world objects when the intro video is playing on the canvas.

Adjusted feedback criteria to be fairer for the player, taking in to account how the player may various seconds to read and understand responses and contemplate their choice.

Updated Topic enum to include HEALTH.